src/VBox/Additions/common/VBoxService/VBoxServiceControlSession.cpp: Fixed warnings found by Parfait (change of signedness on implicit conversion). jiraref:VBP-1424
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107655 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/common/VBoxService/VBoxServiceControlSession.cpp
@@ -668,7 +668,7 @@ static int vgsvcGstCtrlSessionHandleFileReadAt(const PVBOXSERVICECTRLSESSION pSe
             rc = RTFileReadAt(pFile->hFile, (RTFOFF)offReadAt, *ppvScratchBuf, RT_MIN(cbToRead, *pcbScratchBuf), &cbRead);
             if (RT_SUCCESS(rc))
             {
-                offNew = offReadAt + cbRead;
+                offNew = (int64_t)(offReadAt + cbRead);
                 RTFileSeek(pFile->hFile, offNew, RTFILE_SEEK_BEGIN, NULL); /* RTFileReadAt does not always change position. */
             }
             else
